fix: add accessToken check and edit error messages

+++ server/controllers/userCtrl.js
@@ -40,7 +40,6 @@ module.exports.registration = async (req, res) => {
     };
 
     const createdUser = await User.create(preparedForCreate);
-
     res.json({
       ...serialize.serializeAuthUser(createdUser),
       accessToken: tokens.accessToken,
@@ -55,13 +54,11 @@ module.exports.logIn = async (req, res) => {
   try {
     const { username, password } = req.body;
     const foundedUser = await User.findOne({ username });
-    console.log('module.exports.logIn -> foundedUser', foundedUser);
     if (!foundedUser) {
       throw new Error(`Пользователь ${username} не зарегистрирован!`);
     }
 
     const isValidPass = bcrypt.compareSync(password, foundedUser.password);
-
     if (isValidPass) {
       const tokens = await authHelper.createTokens(foundedUser._id);
       const updatedUser = await User.findOneAndUpdate(
@@ -77,10 +74,6 @@ module.exports.logIn = async (req, res) => {
         accessToken: tokens.accessToken,
         accessTokenExpiredAt: tokens.accessTokenExpiredAt,
       });
-      console.log(
-        'module.exports.logIn -> tokens.accessToken',
-        tokens.accessToken
-      );
     } else {
       throw new Error(`Неверные учетные данные!`);
     }
@@ -92,10 +85,7 @@ module.exports.logIn = async (req, res) => {
 module.exports.refreshTokens = async (req, res) => {
   try {
     const refreshToken = req.headers['authorization'];
-    console.log('\n !!!Это рефреш от фронта:', refreshToken);
-
     const foundedUser = await User.findOne({ refreshToken });
-
     if (!foundedUser) {
       throw new Error(`Неправильный refresh токен!`);
     }
@@ -107,7 +97,7 @@ module.exports.refreshTokens = async (req, res) => {
       { new: true }
     );
 
-    res.setHeader('Authorization', tokens.refreshToken);
+    res.setHeader('authorization', tokens.refreshToken);
     res.json(tokens);
   } catch (err) {
     res.status(401).json({ message: err.message });
@@ -120,9 +110,8 @@ module.exports.getProfile = async (req, res) => {
     const result = await jwt.verify(accessToken, secret);
     if (result) {
       const foundedUser = await User.findOne({ _id: result.payload });
-      console.log('module.exports.getProfile -> foundedUser', foundedUser);
       if (!foundedUser) {
-        throw new Error(`Неправильный access токен!`);
+        throw new Error(`Пользователь не найден!`);
       }
 
       res.json(serialize.serializeUser(foundedUser));
@@ -136,41 +125,42 @@ module.exports.changeProfile = async (req, res) => {
   try {
     const accessToken = req.headers['authorization'];
     const result = await jwt.verify(accessToken, secret);
-    if (!req.body.oldPassword) {
-      throw new Error(`Пароль не может быть пустым!`);
-    }
-
-    const oldHash = bcrypt.hashSync(req.body.oldPassword, 10);
-    const newHash = bcrypt.hashSync(req.body.newPassword, 10);
-
-    const foundedUser = await User.findOne({ _id: result.payload });
-    if (!foundedUser) {
-      throw new Error(`Пользователь ${username} не зарегистрирован!`);
-    }
-
-    const isValidPass = bcrypt.compareSync(
-      req.body.oldPassword,
-      foundedUser.password
-    );
-
-    const password = newHash || oldHash;
-    const avatar = req.file
-      ? path.join('/uploads', req.file.filename)
-      : foundedUser.image;
-
-    if (isValidPass && result) {
-      const foundedUser = await User.findOneAndUpdate(
-        { _id: result.payload },
-        { ...req.body, password: password, image: avatar },
-        { new: true }
-      );
+    if (result) {
+      const foundedUser = await User.findOne({ _id: result.payload });
       if (!foundedUser) {
-        throw new Error(`Неправильный access токен или пароль!`);
+        throw new Error(`Пользователь не найден!`);
       }
 
-      res.json(serialize.serializeUser(foundedUser));
-    } else {
-      throw new Error(`Старый пароль неверен!`);
+      if (!req.body.oldPassword) {
+        throw new Error(`Пароль не может быть пустым!`);
+      }
+
+      const oldHash = bcrypt.hashSync(req.body.oldPassword, 10);
+      const newHash = bcrypt.hashSync(req.body.newPassword, 10);
+      const isValidPass = bcrypt.compareSync(
+        req.body.oldPassword,
+        foundedUser.password
+      );
+      const password = newHash || oldHash;
+
+      const avatar = req.file
+        ? path.join('/uploads', req.file.filename)
+        : foundedUser.image;
+
+      if (isValidPass && result) {
+        const foundedUser = await User.findOneAndUpdate(
+          { _id: result.payload },
+          { ...req.body, password: password, image: avatar },
+          { new: true }
+        );
+        if (!foundedUser) {
+          throw new Error(`Ошибка обновления данных!`);
+        }
+
+        res.json(serialize.serializeUser(foundedUser));
+      } else {
+        throw new Error(`Старый пароль неверен!`);
+      }
     }
   } catch (err) {
     if (req.file) {
@@ -181,7 +171,6 @@ module.exports.changeProfile = async (req, res) => {
         }
       );
     }
-
     res.status(401).json({ message: err.message });
   }
 };
@@ -190,11 +179,10 @@ module.exports.getAllUsers = async (req, res) => {
   try {
     const accessToken = req.headers['authorization'];
     const result = await jwt.verify(accessToken, secret);
-
     if (result) {
       const foundedUsers = await User.find();
       if (!foundedUsers) {
-        throw new Error(`Неправильный access токен!`);
+        throw new Error(`Пользователь не найден!`);
       }
 
       const preparedUsers = [];
@@ -210,38 +198,38 @@ module.exports.getAllUsers = async (req, res) => {
 };
 
 module.exports.changeUserPermission = async (req, res) => {
-  console.log('module.exports.changeUserPermission -> req', req.params.id);
   try {
     const accessToken = req.headers['authorization'];
     const result = await jwt.verify(accessToken, secret);
+    if (result) {
+      const foundedUser = await User.findOneAndUpdate(
+        { _id: req.params.id },
+        { ...req.body },
+        { new: true }
+      );
+      if (!foundedUser) {
+        throw new Error(`Пользователь не найден!`);
+      }
 
-    const foundedUser = await User.findOneAndUpdate(
-      { _id: req.params.id },
-      { ...req.body },
-      { new: true }
-    );
-    if (!foundedUser) {
-      throw new Error(`Пользователь ${username} не зарегистрирован!`);
+      res.json(foundedUser.permission);
     }
-
-    res.json(foundedUser.permission);
   } catch (err) {
     res.status(401).json({ message: err.message });
   }
 };
 
 module.exports.deleteUser = async (req, res) => {
-  console.log('module.exports.changeUserPermission -> req', req.params.id);
   try {
     const accessToken = req.headers['authorization'];
     const result = await jwt.verify(accessToken, secret);
+    if (result) {
+      const foundedUser = await User.deleteOne({ _id: req.params.id });
+      if (!foundedUser) {
+        throw new Error(`Пользователь не найден!`);
+      }
 
-    const foundedUser = await User.deleteOne({ _id: req.params.id });
-    if (!foundedUser) {
-      throw new Error(`Пользователь ${username} не зарегистрирован!`);
+      res.json({ message: `Пользователь ${foundedUser.username} удалён!` });
     }
-
-    res.json({ message: `Пользователь ${foundedUser.username} удалён!` });
   } catch (err) {
     res.status(401).json({ message: err.message });
   }
